add schemas test
